libiolog: silence -Wconversion warnings.

+++ include/sudo_iolog.h
@@ -150,6 +150,6 @@ void *iolog_pwfilt_alloc(void);
 bool iolog_pwfilt_add(void *handle, const char *pattern);
 void iolog_pwfilt_free(void *handle);
 bool iolog_pwfilt_remove(void *handle, const char *pattern);
-bool iolog_pwfilt_run(void *handle, int event, const char *buf, unsigned int len, char **newbuf);
+bool iolog_pwfilt_run(void *handle, int event, const char *buf, size_t len, char **newbuf);
 
 #endif /* SUDO_IOLOG_H */
